JS Best Practices applied
